fix: backbone and truth-encoding dtype follow solve policy on MPS
OMAB presence/QKV/out projections and NumericAnswers.encode_targets_batch
still upcast with bare .double(); route through solve_dtype (or the codec's
own dtype) so MPS FP32 runs end to end. CPU/CUDA FP64 reference unchanged.

+++ src/tabu_lab/models/restoration/answers.py
@@ -55,7 +55,7 @@ class NumericAnswers:
             raise ValueError("numeric target values must have floating [column,target] shape")
         finite(values, "numeric target values")
         median, scale = cls._batch_parameters(codecs, values.device)
-        encoded = ((values.detach().double() - median[:, None]) / scale[:, None])[..., None]
+        encoded = ((values.detach().to(median.dtype) - median[:, None]) / scale[:, None])[..., None]
         finite(encoded, "numeric target encoding")
         return encoded
